locking information is not available to the skill.
Context now is made up of two levels:
Context:
\---lock: 'skill_name'
\---data
       \---...
       \---...

The data part is sent to the skill. Since this is a lower level object, the skill has
no way to access the above layer. This change in structure is useful to store
private metadata that the bot doesn't want the skill to be able to access.